Minor changes before collision implementation

+++ game/character.h
@@ -32,8 +32,8 @@ protected:
     int health = 100;
     int accelerationX = 2;
     int acceleartionY = -2;
-    int speedX = 0;
-    int speedY = 0;
+    float speedX = 0;
+    float speedY = 0;
     int maxSpeedX = 20;
     int maxSpeedY = 20;
     int jumpSpeed = 10;
